requirements.txt
Flask==2.2.2
werkzeug==2.2.2
jinja2==3.1.2
gunicorn==20.1.0

psycopg2==2.9.3
SQLAlchemy==1.4.40
SQLAlchemy-Utils==0.38.3
Flask-SQLAlchemy==2.5.1
alembic==1.8.1

redis==4.3.4
celery==5.2.7

pytest==7.1.2
pytest-cov==3.0.0
flake8==5.0.4
black==22.6.0

flask-debugtoolbar==0.13.1
Flask-Static-Digest==0.2.1
Flask-Secrets==0.1.0
Flask-DB==0.3.2

requests
pyyaml